Fixed some bugs
game crash when finishing order: Reason was the inverse trash can outside the truck deleted the whole order when giving it out
Tray carrier of customer was not carrying the tray: Reason was that it was just a marker and because tray is physic object it just fell down.
Recepie detector was reversing the ingredients array every check: Reason was that the order paper needs it reversed but it reversed the original not a copy of it.
Orders with slices could not be finished: Reason was that a slice didn't get the ingredient type of it's origin.